+++ src/platform_impl/linux/wayland/seat/pointer/handlers.rs
@@ -300,17 +300,17 @@ pub(super) fn handle_pointer(
 
 #[inline]
 pub(super) fn handle_relative_pointer(event: RelativePointerEvent, winit_state: &mut WinitState) {
-    match event {
-        RelativePointerEvent::RelativeMotion {
-            dx_unaccel,
-            dy_unaccel,
-            ..
-        } => winit_state.event_sink.push_device_event(
+    if let RelativePointerEvent::RelativeMotion {
+        dx_unaccel,
+        dy_unaccel,
+        ..
+    } = event
+    {
+        winit_state.event_sink.push_device_event(
             DeviceEvent::MouseMotion {
                 delta: (dx_unaccel, dy_unaccel),
             },
             DeviceId,
-        ),
-        _ => (),
+        )
     }
 }